Move core and features into src folder
