+++ lib/src/core/routing/ui/screens/settings_details_screen.dart
