+++ lib/src/core/routing/ui/screens/settings_tab.dart
@@ -1,23 +1,23 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
-import 'package:flutter_boolean_template/core/routing/router/app_router.dart';
+import 'package:flutter_boolean_template/src/core/routing/router/app_router.dart';
 
 @RoutePage()
-class ProfileScreen extends StatelessWidget {
-  const ProfileScreen({super.key});
+class SettingsScreen extends StatelessWidget {
+  const SettingsScreen({super.key});
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Profile'),
+        title: const Text('Settings'),
         leading: const AutoLeadingButton(),
       ),
-      backgroundColor: Colors.deepPurple,
+      backgroundColor: Colors.red,
       body: Center(
         child: FilledButton(
           onPressed: () {
-            context.pushRoute(const ProfileDetailsRoute());
+            context.pushRoute(const SettingsDetailsRoute());
           },
           child: const Text('Push Details'),
         ),
