+++ lib/src/core/store/store_example.dart
