+++ lib/src/features/example_feature/data/dto/weather.dart
