+++ lib/src/features/example_feature/data/repository/weather_repository.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_boolean_template/features/example_feature/data/data_source/weather_api.dart';
-import 'package:flutter_boolean_template/features/example_feature/data/dto/weather.dart';
+import 'package:flutter_boolean_template/src/features/example_feature/data/data_source/weather_api.dart';
+import 'package:flutter_boolean_template/src/features/example_feature/data/dto/weather.dart';
 import 'package:http/http.dart' as http;
 
 class HttpWeatherRepository implements WeatherRepository {
