+++ lib/src/features/example_feature/domain/entity/user_account.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_boolean_template/features/example_feature/domain/value/email.dart';
-import 'package:flutter_boolean_template/features/example_feature/domain/value/id.dart';
-import 'package:flutter_boolean_template/features/example_feature/domain/value/name.dart';
+import 'package:flutter_boolean_template/src/features/example_feature/domain/value/email.dart';
+import 'package:flutter_boolean_template/src/features/example_feature/domain/value/id.dart';
+import 'package:flutter_boolean_template/src/features/example_feature/domain/value/name.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:modddels_annotation_fpdart/modddels_annotation_fpdart.dart';
