+++ lib/src/features/example_feature/domain/entity/user_account.freezed.dart
