+++ lib/src/features/example_feature/domain/value/email.freezed.dart
