+++ lib/src/features/example_feature/domain/value/email.modddel.dart
