+++ lib/src/features/example_feature/domain/value/id.freezed.dart
