+++ lib/src/features/example_feature/domain/value/name.dart
